tema3/requirements.txt
requests~=2.25.1
PyJWT~=1.7.1